+++ src/hooks/index.js
@@ -5,6 +5,8 @@ export { default as useCallout } from './useCallout';
 export { default as useHoldingItemsQuery } from './useHoldingItemsQuery';
 export { default as useHoldingMutation } from './useHoldingMutation';
 export { default as useHoldingsFromStorage } from './useHoldingsFromStorage';
+export { default as useInstanceAuditDataQuery } from './useInstanceAuditDataQuery';
+export { default as useHoldingAuditDataQuery } from './useHoldingAuditDataQuery';
 export { default as useInstanceMutation } from './useInstanceMutation';
 export { default as useHoldingsQueryByHrids } from './useHoldingsQueryByHrids';
 export { default as useInventoryBrowse } from './useInventoryBrowse';
@@ -14,8 +16,7 @@ export { default as useClassificationIdentifierTypes } from './useClassification
 export { default as useClassificationBrowseConfig } from './useClassificationBrowseConfig';
 export { default as useUpdateOwnership } from './useUpdateOwnership';
 export { default as useLocalStorageItems } from './useLocalStorageItems';
-export { default as useHoldingAuditDataQuery } from './useHoldingAuditDataQuery';
-export { default as useVersionHistory } from './useVersionHistory';
+export { default as useInventoryVersionHistory } from './useInventoryVersionHistory';
 export * from './useQuickExport';
 export * from '@folio/stripes-inventory-components/lib/queries/useInstanceDateTypes';
 export * from './useCallNumberTypesQuery';