+++ src/hooks/useInventoryVersionHistory/getActionLabel.test.js
@@ -5,9 +5,9 @@ const intl = { formatMessage: ({ id }) => id };
 describe('getActionLabel', () => {
   it('should return correct action labels', () => {
     const labels = {
-      ADDED: 'stripes-acq-components.audit-log.action.added',
-      MODIFIED: 'stripes-acq-components.audit-log.action.edited',
-      REMOVED: 'stripes-acq-components.audit-log.action.removed',
+      ADDED: 'ui-inventory.versionHistory.action.added',
+      MODIFIED: 'ui-inventory.versionHistory.action.edited',
+      REMOVED: 'ui-inventory.versionHistory.action.removed',
     };
 
     expect(getActionLabel(intl.formatMessage)).toEqual(labels);